Create RestartUtils and wire up all applicable setting
